Split wish command into focused modules
