+++ commands/wish.js
@@ -1,166 +1,24 @@
 'use strict';
 
-const crypto = require('crypto');
-const fs = require('fs');
-const path = require('path');
-const { getLogPath, ensureLogDirectory, createLogFile, addInboxIdea } = require('../lib/file-ops');
-const engineRegistry = require('../lib/engine-registry');
-const { resolveFunctionalOwner } = require('../lib/functional-owner');
-const { buildMissionRoom, writeMissionRoomReceipt } = require('../lib/mission-room');
-const { delegateTask } = require('./task');
-const { startMission, listMissions } = require('./mission');
-
-const VERBS = new Set([
-  'add',
-  'audit',
-  'build',
-  'change',
-  'check',
-  'clean',
-  'create',
-  'debug',
-  'delegate',
-  'deploy',
-  'design',
-  'draft',
-  'explain',
-  'find',
-  'fix',
-  'improve',
-  'make',
-  'migrate',
-  'move',
-  'open',
-  'plan',
-  'polish',
-  'refactor',
-  'remove',
-  'rename',
-  'research',
-  'review',
-  'run',
-  'set',
-  'setup',
-  'ship',
-  'start',
-  'stop',
-  'test',
-  'update',
-  'verify',
-  'write',
-  'understand',
-]);
-
-const PROPER_SKIP = new Set([
-  'A',
-  'An',
-  'And',
-  'At',
-  'But',
-  'For',
-  'I',
-  'In',
-  'Make',
-  'On',
-  'Or',
-  'Please',
-  'The',
-  'This',
-  'To',
-  'With',
-]);
-
-const BUDGET_LABELS = {
-  quick: 'about fifteen minutes',
-  long: 'about an hour',
-  deep: 'about three hours',
-};
-
-const WISH_SWEEP_LIMIT = 3;
-const WISH_MISSION_RUNNER = 'claude';
-const TEST_VERIFY_COMMAND = 'node --test';
-const JOURNAL_RESULT_TEXT = 'a written result will be waiting in your journal';
-const WAITING_INPUT_STATUSES = new Set(['needs_input', 'waiting_input']);
-const AUDIENCE_WORDS = /\b(?:for|user|users|customer|customers|operator|operators|agent|agents|developer|developers|admin|admins|member|members|team|teams|visitor|visitors|client|clients|student|students|founder|founders|newcomer|newcomers)\b/i;
-const BUDGET_RANK = { quick: 1, long: 2, deep: 3 };
-const PART_JOINER = /\s+(?:and|plus|also)\s+|[;]\s*/i;
-const EXTERNAL_REF_WORDS = new Set(['repo', 'repository', 'project', 'workspace']);
-const EXTERNAL_REF_STOP = new Set([
-  'and',
-  'also',
-  'but',
-  'for',
-  'hard',
-  'it',
-  'may',
-  'might',
-  'or',
-  'plus',
-  'seem',
-  'seems',
-  'that',
-  'though',
-  'to',
-  'when',
-  'where',
-  'which',
-  'with',
-]);
-const SUBJECT_SKIP_WORDS = new Set([
-  'can',
-  'could',
-  'done',
-  'everything',
-  'finish',
-  'line',
-  'maybe',
-  'might',
-  'need',
-  'needs',
-  'should',
-  'take',
-  'this',
-  'us',
-  'we',
-  'will',
-  'wish',
-  'would',
-]);
-const FILLER_WORDS = new Set([
-  'a',
-  'an',
-  'and',
-  'are',
-  'as',
-  'at',
-  'be',
-  'better',
-  'by',
-  'for',
-  'from',
-  'good',
-  'great',
-  'help',
-  'improved',
-  'into',
-  'it',
-  'its',
-  'more',
-  'nice',
-  'of',
-  'on',
-  'or',
-  'our',
-  'please',
-  'rough',
-  'roughly',
-  'the',
-  'their',
-  'this',
-  'to',
-  'want',
-  'with',
-]);
+const {
+  analyzeWishParts,
+  auditWish,
+  deriveVerifyPlan,
+  inferBudgetTier,
+  missingNamedInputs,
+  sharesMeaningfulWords,
+  verifyOutcomeText,
+} = require('../lib/wish-audit');
+const { readWishEvents, readWishes, stateFile } = require('../lib/wish-store');
+const {
+  captureWishToJournal,
+  grantWish,
+  printList,
+  reviewWish,
+  runCapturedWish,
+  sweepWishes,
+  waitingOperatorWishes,
+} = require('../lib/wish-delegate');
 
 function showHelp() {
   console.log('');
@@ -171,10 +29,6 @@ function showHelp() {
   console.log('');
 }
 
-function hasFlag(args, name) {
-  return args.includes(name);
-}
-
 function unquote(value) {
   const text = String(value);
   if ((text.startsWith('"') && text.endsWith('"')) || (text.startsWith("'") && text.endsWith("'"))) {
@@ -183,665 +37,37 @@ function unquote(value) {
   return text;
 }
 
-function readFlag(args, name, fallback = '') {
-  const prefix = `${name}=`;
-  for (let i = 0; i < args.length; i += 1) {
-    const arg = String(args[i] || '');
-    if (arg === name && args[i + 1] && !String(args[i + 1]).startsWith('--')) return unquote(args[i + 1]);
-    if (arg.startsWith(prefix)) return unquote(arg.slice(prefix.length));
-  }
-  return fallback;
-}
-
-function stripFlags(args) {
-  const valueFlags = new Set(['--engine']);
-  const out = [];
+function parseFlagArgs(args, valueFlagNames = []) {
+  const valueFlags = new Set(valueFlagNames);
+  const values = {};
+  const positionals = [];
   for (let i = 0; i < args.length; i += 1) {
     const arg = String(args[i] || '');
     if (arg.startsWith('--')) {
       const flagName = arg.split('=')[0];
-      if (valueFlags.has(flagName) && !arg.includes('=') && args[i + 1] && !String(args[i + 1]).startsWith('--')) i += 1;
-      continue;
-    }
-    out.push(args[i]);
-  }
-  return out;
-}
-
-function stripReviewFlags(args) {
-  const valueFlags = new Set(['--score']);
-  const out = [];
-  for (let i = 0; i < args.length; i += 1) {
-    const arg = String(args[i] || '');
-    if (arg.startsWith('--')) {
-      const flagName = arg.split('=')[0];
-      if (valueFlags.has(flagName) && !arg.includes('=') && args[i + 1] && !String(args[i + 1]).startsWith('--')) i += 1;
-      continue;
-    }
-    out.push(args[i]);
-  }
-  return out;
-}
-
-function readEngineOverride(args) {
-  return String(readFlag(args, '--engine', '') || '').trim();
-}
-
-function validateEngineOverride(args, root) {
-  const engineOverride = readEngineOverride(args);
-  if (engineOverride) engineRegistry.resolveRegisteredEngine(engineOverride, root);
-  return engineOverride;
-}
-
-function stateFile(root = process.cwd()) {
-  return path.join(root, '.atris', 'state', 'wishes.jsonl');
-}
-
-function stampIso() {
-  return new Date().toISOString();
-}
-
-function todayName(date = new Date()) {
-  return `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')}`;
-}
-
-function slugify(value) {
-  return String(value || 'wish')
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '')
-    .slice(0, 36) || 'wish';
-}
-
-function wishId(text, ts = stampIso()) {
-  const hash = crypto.createHash('sha1').update(`${text}:${ts}`).digest('hex').slice(0, 8);
-  return `wish-${todayName(new Date(ts))}-${slugify(text)}-${hash}`;
-}
-
-function appendWishRecord(root, record) {
-  const file = stateFile(root);
-  fs.mkdirSync(path.dirname(file), { recursive: true });
-  fs.appendFileSync(file, `${JSON.stringify(record)}\n`, 'utf8');
-  return file;
-}
-
-function readJsonLines(file) {
-  if (!fs.existsSync(file)) return [];
-  return fs.readFileSync(file, 'utf8')
-    .split(/\r?\n/)
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line) => {
-      try {
-        return JSON.parse(line);
-      } catch {
-        return null;
+      if (valueFlags.has(flagName)) {
+        const prefix = flagName + '=';
+        if (arg.startsWith(prefix)) values[flagName] = unquote(arg.slice(prefix.length));
+        else if (args[i + 1] && !String(args[i + 1]).startsWith('--')) {
+          values[flagName] = unquote(args[i + 1]);
+          i += 1;
+        }
       }
-    })
-    .filter(Boolean);
-}
-
-function eventAnswers(event) {
-  const answers = [];
-  for (const key of ['answer', 'answer_text', 'grant_answer']) {
-    const value = event && event[key];
-    if (String(value || '').trim()) answers.push(String(value));
-  }
-  if (Array.isArray(event && event.answers)) {
-    answers.push(...event.answers.map(String).filter((value) => value.trim()));
-  }
-  return answers;
-}
-
-function readWishEvents(root = process.cwd()) {
-  return readJsonLines(stateFile(root));
-}
-
-function readWishes(root = process.cwd()) {
-  const byId = new Map();
-  for (const event of readWishEvents(root)) {
-    if (event && event.kind === 'review') {
-      const wishId = String(event.wish_id || '').trim();
-      if (!wishId) continue;
-      const current = byId.get(wishId) || {
-        id: wishId,
-        ts: event.ts,
-        text: '',
-        answers: [],
-        reviews: [],
-        first_ts: event.ts,
-      };
-      const review = {
-        ts: event.ts,
-        review_text: event.review_text || '',
-        review_score: event.review_score === undefined ? null : event.review_score,
-        reviewed_by: event.reviewed_by || null,
-      };
-      const reviews = [...(current.reviews || []), review];
-      byId.set(wishId, {
-        ...current,
-        reviews,
-        reviewed: true,
-        review_count: reviews.length,
-        latest_review: review,
-      });
       continue;
     }
-    if (!event || !event.id) continue;
-    const current = byId.get(event.id) || {
-      id: event.id,
-      ts: event.ts,
-      text: event.text || '',
-      answers: [],
-      reviews: [],
-      first_ts: event.ts,
-    };
-    const next = {
-      ...current,
-      ...event,
-      first_ts: current.first_ts || event.ts,
-      answers: current.answers || [],
-      reviews: current.reviews || [],
-    };
-    const answers = eventAnswers(event);
-    if (answers.length) next.answers = [...next.answers, ...answers];
-    byId.set(event.id, next);
+    positionals.push(args[i]);
   }
-  return Array.from(byId.values())
-    .sort((a, b) => String(a.first_ts || a.ts || '').localeCompare(String(b.first_ts || b.ts || '')));
+  return { positionals, values };
 }
 
-function latestWishEventMap(root = process.cwd()) {
-  const byId = new Map();
-  for (const event of readWishEvents(root)) {
-    if (event && event.id) byId.set(event.id, event);
-  }
-  return byId;
-}
-
-function captureWishToJournal(text, root = process.cwd()) {
-  if (!fs.existsSync(path.join(root, 'atris'))) {
-    fs.mkdirSync(path.join(root, 'atris'), { recursive: true });
-  }
-  ensureLogDirectory();
-  const { logFile, dateFormatted } = getLogPath();
-  if (!fs.existsSync(logFile)) createLogFile(logFile, dateFormatted);
-  const inboxId = addInboxIdea(logFile, text);
-  return { logFile, inbox_id: inboxId };
-}
-
-function wordList(text) {
-  return String(text || '')
-    .trim()
-    .split(/\s+/)
-    .map((word) => word.trim())
-    .filter(Boolean);
-}
-
-function quoteText(text) {
-  return JSON.stringify(String(text || ''));
-}
-
-function hasVerb(text) {
-  return wordList(text).some((word) => {
-    const clean = word.toLowerCase().replace(/^[^a-z]+|[^a-z]+$/g, '');
-    return VERBS.has(clean) || /(?:ed|ing)$/.test(clean);
-  });
-}
-
-function largerBudget(left, right) {
-  if (!left) return right || null;
-  if (!right) return left;
-  return BUDGET_RANK[right] > BUDGET_RANK[left] ? right : left;
-}
-
-function speedBudgetSignal(text) {
-  const compact = String(text || '').toLowerCase();
-  const action = '(?:fix|patch|polish|tweak|change|update|edit|cleanup|clean|rename|remove|add|copy)';
-  if (new RegExp(`\\b(?:quick|small|tiny)\\s+${action}\\b`).test(compact)) return 'quick';
-  if (new RegExp(`\\b${action}\\b.{0,32}\\b(?:quick|quickly|small|tiny)\\b`).test(compact)) return 'quick';
-  return null;
-}
-
-function deliverableBudgetSignal(text) {
-  const compact = String(text || '').toLowerCase();
-  if (/\b(big|full|overhaul|all|refactor|restructure|rewrite|architecture|system|suite|migration|migrate|e2e|end-to-end)\b/.test(compact)) {
-    return 'deep';
-  }
-  if (/\b(test|tests|testing|coverage|command|commands|cli|subcommand|workflow|flow|integration|real results)\b/.test(compact)) {
-    return 'long';
-  }
-  return null;
-}
-
-function inferBudgetTier(text) {
-  let tier = null;
-  tier = largerBudget(tier, speedBudgetSignal(text));
-  tier = largerBudget(tier, deliverableBudgetSignal(text));
-  return tier || 'long';
-}
-
-function resolveRole(role, root) {
-  const resolver = engineRegistry.resolveRole || engineRegistry.resolve || engineRegistry.resolveEngineForRole;
-  return resolver(role, root);
-}
-
-function resolveExecutor(root, engineOverride = '') {
-  if (engineOverride) {
-    return engineRegistry.resolveEngineForRoleWithPreference('executor', root, engineOverride);
-  }
+function wishOptions(args) {
+  const parsed = parseFlagArgs(args, ['--engine']);
   return {
-    engine: resolveRole('executor', root),
-    requested_engine: null,
-    engine_fallback_reason: null,
+    asJson: args.includes('--json'),
+    noMission: args.includes('--no-mission'),
+    engineOverride: String(parsed.values['--engine'] || '').trim(),
+    positionals: parsed.positionals,
   };
-}
-
-function withRoot(root, fn) {
-  const previous = process.cwd();
-  if (root && previous !== root) process.chdir(root);
-  try {
-    return fn();
-  } finally {
-    if (process.cwd() !== previous) process.chdir(previous);
-  }
-}
-
-function cleanToken(value) {
-  return String(value || '')
-    .replace(/^[`'"]+|[`'",;:!?]+$/g, '')
-    .trim();
-}
-
-function pathLike(value) {
-  const text = cleanToken(value);
-  return text.startsWith('./')
-    || text.startsWith('../')
-    || text.startsWith('/')
-    || text.startsWith('~/')
-    || text.includes('/')
-    || /\.[A-Za-z0-9]{1,8}$/.test(text);
-}
-
-function fileExists(root, value) {
-  const raw = cleanToken(value);
-  if (!raw) return false;
-  const expanded = raw.startsWith('~/') ? path.join(process.env.HOME || '', raw.slice(2)) : raw;
-  const candidates = path.isAbsolute(expanded)
-    ? [expanded]
-    : [path.join(root, expanded), path.join(path.dirname(root), expanded)];
-  return candidates.some((candidate) => fs.existsSync(candidate));
-}
-
-function namedThingExists(root, value) {
-  const raw = cleanToken(value);
-  if (!raw) return false;
-  if (fileExists(root, raw)) return true;
-  const slug = slugify(raw);
-  const base = path.basename(root).toLowerCase();
-  const parentBase = path.basename(path.dirname(root)).toLowerCase();
-  if (slug && (base.includes(slug) || parentBase.includes(slug))) return true;
-  const candidates = [
-    path.join(root, raw),
-    path.join(root, slug),
-    path.join(path.dirname(root), raw),
-    path.join(path.dirname(root), slug),
-    path.join(root, 'atris', 'team', raw),
-    path.join(root, 'atris', 'team', slug),
-  ];
-  return candidates.some((candidate) => fs.existsSync(candidate));
-}
-
-function subjectForWish(text) {
-  const words = wordList(text)
-    .map(cleanToken)
-    .filter(Boolean);
-  if (!words.length) return '';
-  let usable = words;
-  while (usable.length) {
-    const clean = usable[0].toLowerCase().replace(/[^a-z0-9-]/g, '');
-    if (VERBS.has(clean) || FILLER_WORDS.has(clean) || SUBJECT_SKIP_WORDS.has(clean)) {
-      usable = usable.slice(1);
-      continue;
-    }
-    break;
-  }
-  const phrase = [];
-  for (const word of usable) {
-    const clean = word.toLowerCase().replace(/[^a-z0-9-]/g, '');
-    if (!clean) continue;
-    if (['by', 'for', 'from', 'in', 'into', 'of', 'on', 'to', 'with'].includes(clean) && phrase.length) break;
-    if (FILLER_WORDS.has(clean) || SUBJECT_SKIP_WORDS.has(clean)) continue;
-    if (VERBS.has(clean)) continue;
-    phrase.push(clean);
-    if (phrase.length >= 4) break;
-  }
-  return phrase.join(' ');
-}
-
-function meaningfulWordSet(text) {
-  const words = new Set();
-  for (const raw of wordList(text)) {
-    const clean = raw.toLowerCase().replace(/^[^a-z0-9]+|[^a-z0-9]+$/g, '');
-    if (!clean || clean.length < 3) continue;
-    if (VERBS.has(clean) || FILLER_WORDS.has(clean)) continue;
-    words.add(clean.replace(/s$/, ''));
-  }
-  return words;
-}
-
-function sharesMeaningfulWords(left, right) {
-  const leftWords = meaningfulWordSet(left);
-  const rightWords = meaningfulWordSet(right);
-  if (!leftWords.size || !rightWords.size) return false;
-  for (const word of leftWords) {
-    if (rightWords.has(word)) return true;
-  }
-  return false;
-}
-
-function clarityAudit(text) {
-  const words = wordList(text);
-  const subject = subjectForWish(text);
-  const missing = [];
-  const questions = [];
-  const vague = words.length < 4 || !hasVerb(text);
-
-  if (vague) {
-    missing.push('outcome');
-    questions.push(subject ? `What outcome should ${subject} create?` : 'What would done look like for this wish?');
-  }
-  if (vague && !AUDIENCE_WORDS.test(text)) {
-    missing.push('audience');
-    questions.push(subject ? `Who is ${subject} for?` : 'What should exist when it comes true?');
-  }
-  if (vague) {
-    missing.push('scope');
-    questions.push(subject ? `What part of ${subject} should I change first?` : 'What should I change first?');
-  }
-
-  return {
-    vague,
-    missing,
-    questions,
-  };
-}
-
-function missingNamedInputs(text, root) {
-  const missing = [];
-  const seen = new Set();
-  const addMissing = (value, kind) => {
-    const clean = cleanToken(value);
-    if (!clean || seen.has(clean)) return;
-    seen.add(clean);
-    const exists = kind === 'path' ? fileExists(root, clean) : namedThingExists(root, clean);
-    if (!exists) missing.push({ value: clean, kind });
-  };
-
-  const words = wordList(text);
-  for (const word of words) {
-    if (pathLike(word)) addMissing(word, 'path');
-  }
-
-  const keywordPattern = /\b(?:member|repo|repository|file|folder|path)\s+[`'"]?([A-Za-z0-9._/-]+)[`'"]?/gi;
-  let match;
-  while ((match = keywordPattern.exec(text))) {
-    addMissing(match[1], pathLike(match[1]) ? 'path' : 'proper');
-  }
-
-  for (let index = 0; index < words.length; index += 1) {
-    const clean = cleanToken(words[index]);
-    if (!/^[A-Z][A-Za-z0-9_-]{2,}$/.test(clean)) continue;
-    if (PROPER_SKIP.has(clean)) continue;
-    if (index === 0 && VERBS.has(clean.toLowerCase())) continue;
-    addMissing(clean, 'proper');
-  }
-  return missing;
-}
-
-function cleanPartText(text) {
-  return String(text || '')
-    .replace(/\s+/g, ' ')
-    .replace(/^[,;:\s]+|[,;:\s]+$/g, '')
-    .trim();
-}
-
-function externalReference(text) {
-  const words = wordList(text).map(cleanToken).filter(Boolean);
-  for (let index = 0; index < words.length; index += 1) {
-    const marker = words[index].toLowerCase().replace(/[^a-z]/g, '');
-    if (!EXTERNAL_REF_WORDS.has(marker)) continue;
-    const collected = [];
-    for (let next = index + 1; next < words.length; next += 1) {
-      const clean = words[next].toLowerCase().replace(/[^a-z0-9._-]/g, '');
-      if (!clean || EXTERNAL_REF_STOP.has(clean)) break;
-      if (VERBS.has(clean) || FILLER_WORDS.has(clean)) break;
-      collected.push(clean);
-      if (collected.length >= 3) break;
-    }
-    if (collected.length) return { marker, name: collected.join(' ') };
-  }
-  return null;
-}
-
-function outOfScopeReason(text, root) {
-  const ref = externalReference(text);
-  if (ref && !namedThingExists(root, ref.name)) {
-    return `${ref.marker} ${ref.name} is not in this checkout`;
-  }
-  const missing = missingNamedInputs(text, root);
-  const pathMissing = missing.find((item) => item.kind === 'path');
-  if (pathMissing) return `${pathMissing.value} is not in this checkout`;
-  return '';
-}
-
-function partLooksDeliverable(text, root) {
-  if (outOfScopeReason(text, root)) return true;
-  if (hasVerb(text)) return true;
-  return meaningfulWordSet(text).size >= 2;
-}
-
-function analyzeWishParts(text, root) {
-  const clean = cleanPartText(text);
-  if (!clean) return null;
-  const rough = clean.split(PART_JOINER).map(cleanPartText).filter(Boolean);
-  const parts = (rough.length > 1 ? rough : [clean])
-    .map((part, index) => ({
-      part: index + 1,
-      text: part,
-      waiting_reason: outOfScopeReason(part, root),
-    }))
-    .filter((part) => rough.length === 1 || partLooksDeliverable(part.text, root));
-  if (parts.length < 2 && !parts.some((part) => part.waiting_reason)) return null;
-  return parts.length ? parts : null;
-}
-
-function auditWish(text, root = process.cwd(), options = {}) {
-  const questions = [];
-  const clarity = clarityAudit(text);
-  questions.push(...clarity.questions);
-  const engineOverride = String(options.engineOverride || '').trim();
-
-  let executor = null;
-  let validator = null;
-  let requestedEngine = null;
-  let engineFallbackReason = null;
-  try {
-    const resolved = resolveExecutor(root, engineOverride);
-    executor = resolved.engine;
-    requestedEngine = resolved.requested_engine;
-    engineFallbackReason = resolved.engine_fallback_reason;
-  } catch (error) {
-    if (engineOverride) throw error;
-    executor = null;
-  }
-  try {
-    validator = resolveRole('validator', root);
-  } catch {
-    validator = null;
-  }
-  if (!executor) questions.push('Which working builder should handle this?');
-  if (!validator) questions.push('Which working reviewer should validate it?');
-
-  const missing = missingNamedInputs(text, root);
-  if (missing.some((item) => item.kind === 'path')) {
-    questions.push('Where should I find the file or folder you named?');
-  } else if (missing.length) {
-    questions.push(`Which workspace, repo, file, or team member did you mean by ${missing[0].value}?`);
-  }
-
-  return {
-    ok: questions.length === 0,
-    questions: questions.slice(0, 3),
-    executor,
-    validator,
-    budget: inferBudgetTier(text),
-    missing,
-    vague: clarity.vague,
-    missing_slots: clarity.missing,
-    requested_engine: requestedEngine,
-    engine_fallback_reason: engineFallbackReason,
-  };
-}
-
-function machineRecord(wish, status, audit, extra = {}) {
-  const record = {
-    wish_id: wish.id,
-    status,
-    task_id: extra.task_id || null,
-    mission_id: extra.mission_id || null,
-    engine: extra.engine || (audit && audit.executor ? audit.executor.id : null),
-    budget: audit ? audit.budget : (extra.budget || null),
-    questions: audit ? audit.questions : (extra.questions || []),
-  };
-  const requestedEngine = extra.requested_engine || (audit && audit.requested_engine);
-  const engineFallbackReason = extra.engine_fallback_reason || (audit && audit.engine_fallback_reason);
-  if (requestedEngine) record.requested_engine = requestedEngine;
-  if (engineFallbackReason) record.engine_fallback_reason = engineFallbackReason;
-  return record;
-}
-
-function engineAuditFields(audit) {
-  const fields = {};
-  if (audit && audit.requested_engine) fields.requested_engine = audit.requested_engine;
-  if (audit && audit.engine_fallback_reason) fields.engine_fallback_reason = audit.engine_fallback_reason;
-  return fields;
-}
-
-function wishDelegationEngine(audit, options = {}) {
-  if (options.engine) return String(options.engine);
-  if (audit && audit.requested_engine && audit.executor && audit.executor.id) return audit.executor.id;
-  return WISH_MISSION_RUNNER;
-}
-
-function printQuestions(wishText, questions) {
-  console.log(`You wished: ${quoteText(wishText)}`);
-  questions.forEach((question, index) => {
-    console.log(`${index + 1}. ${question}`);
-  });
-  console.log('answer with atris wish grant <n> "your answer".');
-}
-
-function engineLabel(engine) {
-  const text = String(engine || 'the worker');
-  if (text === 'codex') return 'Codex';
-  if (text === 'claude') return 'Claude';
-  if (text === 'atris-fast') return 'Atris Fast';
-  return text;
-}
-
-function deriveVerifyPlan(text) {
-  const compact = String(text || '').toLowerCase();
-  if (/\b(test|tests|testing|suite|coverage)\b/.test(compact)) {
-    return {
-      command: TEST_VERIFY_COMMAND,
-      outcome: 'the fast test run passes and is timed',
-      status: 'derived',
-    };
-  }
-  if (/\b(command|commands|cli|subcommand)\b/.test(compact)) {
-    return {
-      command: TEST_VERIFY_COMMAND,
-      outcome: 'the new command runs end to end',
-      status: 'derived',
-    };
-  }
-  return {
-    command: '',
-    outcome: 'I will show you the result to judge',
-    status: 'needs-review',
-  };
-}
-
-function verifyOutcomeText(planOrCommand) {
-  if (planOrCommand && typeof planOrCommand === 'object') return planOrCommand.outcome || JOURNAL_RESULT_TEXT;
-  const verify = String(planOrCommand || '').trim();
-  if (!verify) return JOURNAL_RESULT_TEXT;
-  if (verify === TEST_VERIFY_COMMAND) return 'the fast test run passes and is timed';
-  return `the verify command ${quoteText(verify)} passes`;
-}
-
-function printGranted(text, audit, options = {}) {
-  const verifyPlan = options.verifyPlan || deriveVerifyPlan(text);
-  if (options.grantNumber) {
-    console.log(`Granting wish ${options.grantNumber}: ${quoteText(text)}`);
-  } else {
-    console.log(`I heard you: ${quoteText(text)}`);
-  }
-  console.log(`I delegated it to ${engineLabel(options.engine || audit.executor.id)} with a ${audit.budget} budget, ${BUDGET_LABELS[audit.budget]}.`);
-  if (verifyPlan.status === 'needs-review') console.log(verifyPlan.outcome + '.');
-  else console.log(`You will know it came true when ${verifyOutcomeText(verifyPlan)}.`);
-}
-
-function latestMissionStatus(root, missionId) {
-  if (!missionId) return '';
-  try {
-    const mission = listMissions(root).find((row) => row.id === missionId);
-    return mission ? String(mission.status || '') : '';
-  } catch {
-    return '';
-  }
-}
-
-function operatorStatus(wish, root = process.cwd()) {
-  if (WAITING_INPUT_STATUSES.has(String(wish.status || ''))) return 'waiting on you';
-  if (wish.status === 'delegated' || wish.status === 'decomposed') {
-    const missionStatus = latestMissionStatus(root, wish.mission_id);
-    if (missionStatus === 'complete') return 'came true';
-    if (missionStatus === 'stopped') return 'stopped';
-    if (missionStatus === 'blocked') return 'blocked';
-    if (missionStatus === 'paused') return 'paused';
-    if (missionStatus === 'ready') return 'ready for review';
-    if (missionStatus) return 'in flight';
-    if (Array.isArray(wish.out_of_scope_parts) && wish.out_of_scope_parts.length) return 'waiting on another home';
-    return 'in flight';
-  }
-  if (wish.status === 'complete') return 'came true';
-  return wish.status || 'waiting';
-}
-
-function reviewedMarker(wish) {
-  return wish && (wish.reviewed || (Array.isArray(wish.reviews) && wish.reviews.length)) ? ' [reviewed]' : '';
-}
-
-function openWishes(root = process.cwd()) {
-  return readWishes(root).filter((wish) => ['needs_input', 'waiting_input', 'delegated', 'decomposed', 'complete'].includes(String(wish.status || '')));
-}
-
-function printList(root = process.cwd()) {
-  const wishes = openWishes(root);
-  if (!wishes.length) {
-    console.log('No open wishes.');
-    return 0;
-  }
-  wishes.forEach((wish, index) => {
-    console.log(`${index + 1}. ${wish.text} - ${operatorStatus(wish, root)}${reviewedMarker(wish)}`);
-  });
-  return 0;
-}
-
-function actorName() {
-  return process.env.ATRIS_AGENT_ID || process.env.USER || 'operator';
 }
 
 function parseReviewScore(args) {
@@ -850,7 +76,8 @@ function parseReviewScore(args) {
     return text === '--score' || text.startsWith('--score=');
   });
   if (!hasScore) return { ok: true, value: null };
-  const raw = String(readFlag(args, '--score', '') || '').trim();
+  const parsed = parseFlagArgs(args, ['--score']);
+  const raw = String(parsed.values['--score'] || '').trim();
   const allowed = new Set([-1, 0, 1, 2, 3, 4, 5]);
   const value = Number(raw);
   if (!raw || !Number.isInteger(value) || !allowed.has(value)) {
@@ -862,619 +89,12 @@ function parseReviewScore(args) {
   return { ok: true, value };
 }
 
-function reviewableWish(wish) {
-  const status = String(wish && wish.status || '').trim();
-  return status === 'delegated' || status === 'complete' || status === 'completed';
-}
-
-function wishReviewTime(wish, fallback) {
-  const parsed = Date.parse(wish.completed_at || wish.dispatched_at || wish.ts || wish.first_ts || '');
-  return Number.isFinite(parsed) ? parsed : fallback;
-}
-
-function latestReviewableWish(wishes) {
-  let latest = null;
-  let latestTime = -Infinity;
-  wishes.forEach((wish, index) => {
-    if (!reviewableWish(wish)) return;
-    const time = wishReviewTime(wish, index);
-    if (time >= latestTime) {
-      latest = wish;
-      latestTime = time;
-    }
-  });
-  return latest;
-}
-
-function looksLikeWishId(value) {
-  return /^wish[-_]/i.test(String(value || '').trim());
-}
-
-function resolveReviewTarget(root, ref) {
-  const wishes = readWishes(root);
-  if (!wishes.length) {
-    return { ok: false, message: 'No wishes to review yet.' };
-  }
-  if (!ref || ref === 'latest') {
-    const latest = latestReviewableWish(wishes);
-    if (!latest) return { ok: false, message: 'No delegated or completed wishes to review.' };
-    return { ok: true, wish: latest };
-  }
-  const wish = wishes.find((row) => row.id === ref);
-  if (!wish) return { ok: false, message: `No wish found with id ${ref}.` };
-  return { ok: true, wish };
-}
-
-function reviewWish(args, root = process.cwd()) {
-  const score = parseReviewScore(args);
-  if (!score.ok) {
-    console.error(score.message);
-    return 2;
-  }
-  const positionals = stripReviewFlags(args);
-  const parts = positionals.slice(1).map(String).filter((value) => value.trim());
-  if (!parts.length) {
-    console.error('wish review needs one sentence of feedback.');
-    return 2;
-  }
-  const knownIds = new Set(readWishes(root).map((wish) => wish.id));
-  let targetRef = 'latest';
-  let reviewParts = parts;
-  const first = parts[0];
-  if (first === 'latest' || knownIds.has(first) || looksLikeWishId(first)) {
-    targetRef = first;
-    reviewParts = parts.slice(1);
-  }
-  const reviewText = reviewParts.join(' ').trim();
-  if (!reviewText) {
-    console.error('wish review needs one sentence of feedback.');
-    return 2;
-  }
-  const target = resolveReviewTarget(root, targetRef);
-  if (!target.ok) {
-    console.error(target.message);
-    return 2;
-  }
-  appendWishRecord(root, {
-    kind: 'review',
-    wish_id: target.wish.id,
-    ts: stampIso(),
-    review_text: reviewText,
-    review_score: score.value,
-    reviewed_by: actorName(),
-  });
-  console.log(`Review captured for ${quoteText(target.wish.text || target.wish.id)}.`);
-  return 0;
-}
-
-function startWishDelegation(wish, audit, root, options = {}) {
-  const taskText = String(options.taskText || wish.text);
-  const recordText = String(options.recordText || wish.text);
-  const verifyPlan = options.verifyPlan || deriveVerifyPlan(taskText);
-  const engine = wishDelegationEngine(audit, options);
-  const now = stampIso();
-  const ownerResolution = resolveFunctionalOwner({
-    requestedOwner: audit.executor.id,
-    title: taskText,
-    tag: 'wish',
-    goal: taskText,
-    root,
-    fallbackOwners: ['mission-lead', 'task-planner', 'architect', 'validator'],
-  });
-  const baseRoom = buildMissionRoom(taskText, {
-    owner: ownerResolution.owner,
-    root,
-    ownerResolution,
-    trustedRun: true,
-    verifier: verifyPlan.command || '',
-  });
-  const room = {
-    ...baseRoom,
-    wish_id: wish.id,
-    wish: {
-      id: wish.id,
-      text: recordText,
-      task_text: taskText,
-    },
-    source: {
-      ...(baseRoom.source || {}),
-      wish_id: wish.id,
-      wish_text: recordText,
-    },
-  };
-  const writtenRoom = writeMissionRoomReceipt(room, { root });
-  const routeOwner = writtenRoom.room?.member_route?.suggested_member
-    || writtenRoom.room?.owner
-    || ownerResolution.owner
-    || 'mission-lead';
-  const note = [
-    `Wish: ${taskText}`,
-    `Wish id: ${wish.id}`,
-    `Mission Room: ${writtenRoom.relativePath}`,
-    verifyPlan.command
-      ? `Verify: ${verifyPlan.command} (${verifyOutcomeText(verifyPlan)})`
-      : `Verify: needs human review (${verifyPlan.outcome})`,
-    `Budget: ${audit.budget}`,
-  ].join('\n');
-  const { taskPayload, missionPayload } = withRoot(root, () => {
-    const delegated = delegateTask([
-      taskText,
-      '--to',
-      routeOwner,
-      '--executed-by',
-      engine,
-      '--tag',
-      'wish',
-      '--goal-objective',
-      taskText,
-      '--note',
-      note,
-    ]);
-    const missionArgs = [
-      taskText,
-      '--owner',
-      routeOwner,
-      '--runner',
-      engine,
-      '--budget',
-      audit.budget,
-      '--task',
-      delegated.task_id,
-      '--duplicate',
-      '--json',
-    ];
-    if (verifyPlan.command) missionArgs.push('--verify', verifyPlan.command);
-    return {
-      taskPayload: delegated,
-      missionPayload: startMission(missionArgs, {
-        silent: true,
-        missionPatch: {
-          wish_id: wish.id,
-          wish_text: recordText,
-          mission_room_receipt_path: writtenRoom.relativePath,
-          mission_room_name: writtenRoom.room?.name || null,
-          source: 'wish',
-          metadata: {
-            wish_id: wish.id,
-            wish_text: recordText,
-            mission_room_receipt_path: writtenRoom.relativePath,
-            mission_room_name: writtenRoom.room?.name || null,
-            mission_room_owner: routeOwner,
-          },
-        },
-      }),
-    };
-  });
-  const mission = missionPayload && missionPayload.mission ? missionPayload.mission : null;
-  const record = {
-    id: wish.id,
-    ts: now,
-    text: recordText,
-    status: 'delegated',
-    dispatched_at: now,
-    task_id: taskPayload.task_id,
-    mission_id: mission ? mission.id : null,
-    engine,
-    validator: audit.validator.id,
-    budget: audit.budget,
-    verify: verifyPlan.command,
-    verify_status: verifyPlan.status,
-    verify_outcome: verifyPlan.outcome,
-    task_text: taskText,
-    mission_room_receipt_path: writtenRoom.relativePath,
-    mission_room_name: writtenRoom.room?.name || null,
-    mission_owner: routeOwner,
-    ...engineAuditFields(audit),
-  };
-  return { record, taskPayload, mission, verifyPlan };
-}
-
-function appendDelegatedWishRecord(wish, audit, root, options = {}) {
-  const { record, verifyPlan } = startWishDelegation(wish, audit, root, options);
-  appendWishRecord(root, record);
-  return { record, verifyPlan };
-}
-
-function delegateWish(wish, audit, root, asJson, options = {}) {
-  const { record, verifyPlan } = appendDelegatedWishRecord(wish, audit, root, options);
-  const payload = machineRecord(wish, 'delegated', audit, {
-    task_id: record.task_id,
-    mission_id: record.mission_id,
-    engine: record.engine,
-  });
-  if (asJson) console.log(JSON.stringify(payload, null, 2));
-  else printGranted(options.taskText || wish.text, audit, { ...options, verifyPlan, engine: record.engine });
-  return 0;
-}
-
-function captureOnlyWish(wish, audit, root, asJson) {
-  const now = stampIso();
-  appendWishRecord(root, {
-    id: wish.id,
-    ts: now,
-    text: wish.text,
-    status: 'captured_no_mission',
-    no_mission: true,
-    budget: audit ? audit.budget : inferBudgetTier(wish.text),
-    questions: audit ? audit.questions : [],
-    ...engineAuditFields(audit),
-  });
-  const payload = machineRecord(wish, 'captured', audit, {
-    engine: null,
-    budget: audit ? audit.budget : inferBudgetTier(wish.text),
-  });
-  if (asJson) console.log(JSON.stringify(payload, null, 2));
-  else {
-    console.log(`I heard you: ${quoteText(wish.text)}`);
-    console.log('Captured only; no mission started.');
-  }
-  return 0;
-}
-
-function formatPartRefs(numbers) {
-  const sorted = [...numbers].sort((a, b) => a - b);
-  if (!sorted.length) return 'no parts';
-  if (sorted.length === 1) return `part ${sorted[0]}`;
-  const contiguous = sorted.every((value, index) => index === 0 || value === sorted[index - 1] + 1);
-  if (contiguous) return `parts ${sorted[0]}-${sorted[sorted.length - 1]}`;
-  return `parts ${sorted.slice(0, -1).join(', ')} and ${sorted[sorted.length - 1]}`;
-}
-
-function partAgreement(numbers, singular, plural) {
-  return numbers.length === 1 ? singular : plural;
-}
-
-function printDecomposed(parts, delegatedParts, waitingParts) {
-  console.log(`This wish has ${parts.length} ${parts.length === 1 ? 'part' : 'parts'}.`);
-  parts.forEach((part) => {
-    console.log(`Part ${part.part}: ${part.text}`);
-  });
-  const delegatedNumbers = delegatedParts.map((part) => part.part);
-  const ownHomeNumbers = waitingParts
-    .filter((part) => part.reason && /not in this checkout/.test(part.reason))
-    .map((part) => part.part);
-  const otherWaitingNumbers = waitingParts
-    .filter((part) => !ownHomeNumbers.includes(part.part))
-    .map((part) => part.part);
-  if (delegatedNumbers.length && ownHomeNumbers.length) {
-    const needs = partAgreement(ownHomeNumbers, 'needs its own home', 'need their own homes');
-    console.log(`I can start ${formatPartRefs(delegatedNumbers)} now; ${formatPartRefs(ownHomeNumbers)} ${needs}.`);
-  } else if (delegatedNumbers.length) {
-    console.log(`I can start ${formatPartRefs(delegatedNumbers)} now.`);
-  } else if (ownHomeNumbers.length) {
-    const needs = partAgreement(ownHomeNumbers, 'needs its own home', 'need their own homes');
-    console.log(`${formatPartRefs(ownHomeNumbers)} ${needs}.`);
-  }
-  if (otherWaitingNumbers.length) {
-    const need = partAgreement(otherWaitingNumbers, 'needs', 'need');
-    console.log(`${formatPartRefs(otherWaitingNumbers)} ${need} one clearer answer before I start.`);
-  }
-}
-
-function decomposeWish(wish, parts, root, asJson, options = {}) {
-  const delegatedParts = [];
-  const waitingParts = [];
-  let firstAudit = null;
-  for (const part of parts) {
-    if (part.waiting_reason) {
-      waitingParts.push({
-        part: part.part,
-        text: part.text,
-        status: 'waiting',
-        reason: part.waiting_reason,
-      });
-      continue;
-    }
-    const audit = auditWish(part.text, root, options);
-    if (!audit.ok) {
-      waitingParts.push({
-        part: part.part,
-        text: part.text,
-        status: 'waiting',
-        reason: audit.questions[0] || 'needs a clearer scope',
-        questions: audit.questions,
-      });
-      continue;
-    }
-    if (!firstAudit) firstAudit = audit;
-    const { record, verifyPlan } = startWishDelegation(wish, audit, root, {
-      taskText: part.text,
-      recordText: wish.text,
-      verifyPlan: deriveVerifyPlan(part.text),
-    });
-    const delegated = {
-      part: part.part,
-      text: part.text,
-      status: 'delegated',
-      task_id: record.task_id,
-      mission_id: record.mission_id,
-      dispatched_at: record.dispatched_at,
-      budget: record.budget,
-      engine: record.engine,
-      verify: record.verify,
-      verify_status: record.verify_status,
-      verify_outcome: record.verify_outcome,
-      ...engineAuditFields(audit),
-    };
-    delegatedParts.push(delegated);
-    appendWishRecord(root, {
-      ...record,
-      decomposed_part: part.part,
-      parts_total: parts.length,
-      verify_status: verifyPlan.status,
-    });
-  }
-  const statusParts = parts.map((part) => {
-    const delegated = delegatedParts.find((item) => item.part === part.part);
-    if (delegated) return { part: part.part, text: part.text, status: 'delegated' };
-    const waiting = waitingParts.find((item) => item.part === part.part);
-    return {
-      part: part.part,
-      text: part.text,
-      status: 'waiting',
-      reason: waiting ? waiting.reason : 'needs its own home',
-    };
-  });
-  const firstDelegated = delegatedParts[0] || null;
-  const record = {
-    id: wish.id,
-    ts: stampIso(),
-    text: wish.text,
-    status: 'decomposed',
-    task_id: firstDelegated ? firstDelegated.task_id : null,
-    mission_id: firstDelegated ? firstDelegated.mission_id : null,
-    engine: firstAudit && firstAudit.executor ? firstAudit.executor.id : null,
-    validator: firstAudit && firstAudit.validator ? firstAudit.validator.id : null,
-    budget: firstAudit ? firstAudit.budget : inferBudgetTier(wish.text),
-    parts: statusParts,
-    delegated_parts: delegatedParts,
-    ...engineAuditFields(firstAudit),
-    out_of_scope_parts: waitingParts
-      .filter((part) => part.reason && /not in this checkout/.test(part.reason))
-      .map((part) => ({ ...part, status: 'waiting' })),
-    waiting_parts: waitingParts,
-  };
-  appendWishRecord(root, record);
-  const payload = {
-    ...machineRecord(wish, 'decomposed', firstAudit, {
-      task_id: record.task_id,
-      mission_id: record.mission_id,
-      engine: record.engine,
-      budget: record.budget,
-      questions: [],
-    }),
-    parts: statusParts,
-    delegated_parts: delegatedParts,
-    out_of_scope_parts: record.out_of_scope_parts,
-  };
-  if (asJson) console.log(JSON.stringify(payload, null, 2));
-  else printDecomposed(parts, delegatedParts, waitingParts);
-  return delegatedParts.length ? 0 : 1;
-}
-
-function askForInput(wish, audit, root, asJson) {
-  appendNeedsInputRecord(wish, audit, root);
-  const payload = machineRecord(wish, 'needs_input', audit);
-  if (asJson) console.log(JSON.stringify(payload, null, 2));
-  else printQuestions(wish.text, audit.questions);
-  return 1;
-}
-
-function appendNeedsInputRecord(wish, audit, root) {
-  appendWishRecord(root, {
-    id: wish.id,
-    ts: stampIso(),
-    text: wish.text,
-    status: 'needs_input',
-    questions: audit.questions,
-    vague: !!audit.vague,
-    missing_slots: audit.missing_slots || [],
-    ...engineAuditFields(audit),
-  });
-}
-
-function latestStatusFor(wish, latestEvent) {
-  return String((latestEvent && latestEvent.status) || wish.status || '').trim();
-}
-
-function latestEventHasGrantAnswer(latestEvent) {
-  return eventAnswers(latestEvent).length > 0;
-}
-
-function dispatchableWishReason(wish, latestEvent) {
-  if (!wish || !wish.id || wish.dispatched_at || wish.no_mission || (latestEvent && latestEvent.no_mission)) return '';
-  const status = latestStatusFor(wish, latestEvent);
-  if (status === 'captured') return 'captured';
-  if (WAITING_INPUT_STATUSES.has(status) && latestEventHasGrantAnswer(latestEvent)) return 'answered';
-  return '';
-}
-
-function waitingWishNeed(questions = []) {
-  const first = (Array.isArray(questions) ? questions : [])
-    .map((question) => String(question || '').trim())
-    .find(Boolean) || 'Answer the open question before I dispatch it.';
-  if (/working builder/i.test(first)) return 'needs a working builder before it can start';
-  if (/working reviewer/i.test(first)) return 'needs a working reviewer before it can start';
-  if (/file or folder/i.test(first)) return 'needs the file or folder location before it can start';
-  if (/workspace, repo, file, or team member/i.test(first)) return first.replace(/^Which /, 'needs the ');
-  return first;
-}
-
-function waitingOperatorWishes(root = process.cwd()) {
-  const latest = latestWishEventMap(root);
-  return readWishes(root)
-    .filter((wish) => {
-      const latestEvent = latest.get(wish.id) || wish;
-      const status = latestStatusFor(wish, latestEvent);
-      return WAITING_INPUT_STATUSES.has(status) && !latestEventHasGrantAnswer(latestEvent) && !wish.dispatched_at;
-    })
-    .map((wish) => {
-      const latestEvent = latest.get(wish.id) || wish;
-      const questions = Array.isArray(latestEvent.questions) ? latestEvent.questions : (wish.questions || []);
-      return {
-        id: wish.id,
-        text: String(wish.text || '').trim(),
-        questions,
-        need: waitingWishNeed(questions),
-        first_ts: wish.first_ts || wish.ts || latestEvent.ts || null,
-      };
-    });
-}
-
-function dispatchWishHeadlessly(wish, root, reason) {
-  const auditText = reason === 'answered'
-    ? [wish.text, ...(wish.answers || [])].join(' ')
-    : wish.text;
-  const audit = auditWish(auditText, root);
-  if (!audit.ok) {
-    appendNeedsInputRecord(wish, audit, root);
-    return {
-      id: wish.id,
-      status: 'waiting_on_operator',
-      no_executor: !audit.executor,
-      no_validator: !audit.validator,
-      questions: audit.questions,
-    };
-  }
-  const { record } = appendDelegatedWishRecord(wish, audit, root);
+function reviewOptions(args) {
+  const parsed = parseFlagArgs(args, ['--score']);
   return {
-    id: wish.id,
-    status: 'delegated',
-    dispatched: true,
-    task_id: record.task_id,
-    mission_id: record.mission_id,
-    engine: record.engine,
-    dispatched_at: record.dispatched_at,
+    score: parseReviewScore(args),
+    positionals: parsed.positionals,
   };
-}
-
-function sweepWishes(root = process.cwd(), options = {}) {
-  const limit = Number.isInteger(options.limit) && options.limit >= 0 ? options.limit : WISH_SWEEP_LIMIT;
-  const latest = latestWishEventMap(root);
-  const summary = {
-    scanned: 0,
-    dispatched: 0,
-    capped: 0,
-    waiting_on_operator: 0,
-    skipped_no_executor: 0,
-    results: [],
-  };
-  for (const wish of readWishes(root)) {
-    const latestEvent = latest.get(wish.id) || wish;
-    const reason = dispatchableWishReason(wish, latestEvent);
-    if (!reason) continue;
-    summary.scanned += 1;
-    if (summary.dispatched >= limit) {
-      summary.capped += 1;
-      continue;
-    }
-    const result = dispatchWishHeadlessly(wish, root, reason);
-    summary.results.push(result);
-    if (result.dispatched) summary.dispatched += 1;
-    if (result.no_executor) summary.skipped_no_executor += 1;
-  }
-  summary.waiting_on_operator = waitingOperatorWishes(root).length;
-  return summary;
-}
-
-function runCapturedWish(text, args, root = process.cwd()) {
-  const asJson = hasFlag(args, '--json');
-  const noMission = hasFlag(args, '--no-mission');
-  let engineOverride = '';
-  try {
-    engineOverride = validateEngineOverride(args, root);
-  } catch (error) {
-    console.error(error.message || String(error));
-    return 2;
-  }
-  const ts = stampIso();
-  const wish = {
-    id: wishId(text, ts),
-    ts,
-    text,
-    status: 'captured',
-  };
-  captureWishToJournal(text, root);
-  appendWishRecord(root, wish);
-  const parts = analyzeWishParts(text, root);
-  if (parts && noMission) return captureOnlyWish(wish, null, root, asJson);
-  if (parts) return decomposeWish(wish, parts, root, asJson, { engineOverride });
-  const audit = auditWish(text, root, { engineOverride });
-  if (!audit.ok) return askForInput(wish, audit, root, asJson);
-  if (noMission) return captureOnlyWish(wish, audit, root, asJson);
-  return delegateWish(wish, audit, root, asJson);
-}
-
-function grantWish(args, root = process.cwd()) {
-  const asJson = hasFlag(args, '--json');
-  const noMission = hasFlag(args, '--no-mission');
-  let engineOverride = '';
-  try {
-    engineOverride = validateEngineOverride(args, root);
-  } catch (error) {
-    console.error(error.message || String(error));
-    return 2;
-  }
-  const positionals = stripFlags(args);
-  const number = Number(positionals[1]);
-  const answer = positionals.slice(2).join(' ').trim();
-  if (!Number.isInteger(number) || number <= 0 || !answer) {
-    console.error('wish grant needs a list number and an answer.');
-    return 2;
-  }
-  const waiting = openWishes(root);
-  const wish = waiting[number - 1];
-  if (!wish) {
-    console.error('No wish is waiting at that number.');
-    return 2;
-  }
-  if (!WAITING_INPUT_STATUSES.has(String(wish.status || ''))) {
-    if (!asJson) console.log(`Granting wish ${number}: ${quoteText(wish.text)}`);
-    console.error('That wish is not waiting on an answer.');
-    return 2;
-  }
-  const vagueFlagged = wish.vague === true
-    || (Array.isArray(wish.missing_slots) && wish.missing_slots.length > 0)
-    || (Array.isArray(wish.questions) && wish.questions.some((question) => /\b(outcome|Who is|What part)\b/i.test(String(question))));
-  if (vagueFlagged && !sharesMeaningfulWords(wish.text, answer)) {
-    const notice = 'This answer may be for a different wish, so I did not dispatch it.';
-    if (asJson) {
-      const payload = machineRecord(wish, 'needs_input', {
-        executor: null,
-        budget: wish.budget || inferBudgetTier(wish.text),
-        questions: wish.questions || [],
-      }, {
-        engine: wish.engine || null,
-      });
-      console.log(JSON.stringify({
-        ...payload,
-        mismatch: true,
-        notice,
-        wish_text: wish.text,
-      }, null, 2));
-    } else {
-      console.log(`Granting wish ${number}: ${quoteText(wish.text)}`);
-      console.log(notice);
-      printList(root);
-    }
-    return 1;
-  }
-  appendWishRecord(root, {
-    id: wish.id,
-    ts: stampIso(),
-    text: wish.text,
-    status: 'needs_input',
-    answer,
-  });
-  const answeredWish = {
-    ...wish,
-    answers: [...(wish.answers || []), answer],
-  };
-  const auditText = [wish.text, ...(answeredWish.answers || [])].join(' ');
-  const audit = auditWish(auditText, root, { engineOverride });
-  if (!audit.ok) return askForInput(answeredWish, audit, root, asJson);
-  if (noMission) return captureOnlyWish(answeredWish, audit, root, asJson);
-  return delegateWish(answeredWish, audit, root, asJson, { grantNumber: number });
 }
 
 function wishCommand(args = []) {
@@ -1487,17 +107,24 @@ function wishCommand(args = []) {
     return printList(process.cwd());
   }
   if (first === 'grant' || first === 'answer') {
-    return grantWish(args, process.cwd());
+    const options = wishOptions(args);
+    return grantWish(options.positionals, process.cwd(), options);
   }
   if (first === 'review') {
-    return reviewWish(args, process.cwd());
+    const options = reviewOptions(args);
+    if (!options.score.ok) {
+      console.error(options.score.message);
+      return 2;
+    }
+    return reviewWish(options.positionals, process.cwd(), { reviewScore: options.score.value });
   }
-  const text = stripFlags(args).join(' ').trim();
+  const options = wishOptions(args);
+  const text = options.positionals.join(' ').trim();
   if (!text) {
     showHelp();
     return 2;
   }
-  return runCapturedWish(text, args, process.cwd());
+  return runCapturedWish(text, process.cwd(), options);
 }
 
 module.exports = {
